+++ package.json
@@ -12,8 +12,10 @@
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@fortawesome/vue-fontawesome": "^2.0.2",
     "core-js": "^3.6.5",
+    "date-fns": "^2.22.1",
     "dedent": "^0.7.0",
     "firebase": "^8.6.3",
+    "moment": "^2.29.1",
     "vue": "^2.6.11",
     "vue-quill-editor": "^3.0.6",
     "vue-router": "^3.2.0",
